+++ OTA_server.h
@@ -3,7 +3,7 @@
 #include <ESPmDNS.h>
 #include <Update.h>
 #include <EEPROM.h>
-#include "Rtos5.h"
+#include "ESP_Mini.h"
 #include "OTA_html.h"
 bool downloading_file=false;
 const char* host = "esp32";
@@ -374,17 +374,7 @@ void handleConfigUpload() {
     doc["cal_speed"] = serialized(server.arg("cal_speed")); 
     doc["sample_rate"] = server.arg("sample_rate").toInt();
     doc["gnss"] = server.arg("gnss").toInt();
-    doc["speed_field"] = server.arg("speed_field").toInt();
-    doc["speed_large_font"] = server.arg("speed_large_font").toInt();
-    doc["bar_length"] = server.arg("bar_length").toInt();
-    doc["Stat_screens"] = server.arg("Stat_screens").toInt(); 
-    doc["Stat_screens_time"] = server.arg("Stat_screens_time").toInt(); 
-    doc["stat_speed"] = server.arg("stat_speed").toInt(); 
     doc["archive_days"] = server.arg("archive_days").toInt(); 
-    doc["GPIO12_screens"] = server.arg("GPIO12_screens").toInt(); 
-    doc["Board_Logo"] = server.arg("Board_Logo").toInt();
-    doc["Sail_Logo"] = server.arg("Sail_Logo").toInt();
-    doc["sleep_off_screen"] = server.arg("sleep_off_screen").toInt();
     doc["logTXT"] = server.arg("logTXT").toInt(); 
     doc["logSBP"] = server.arg("logSBP").toInt(); 
     doc["logUBX"] = server.arg("logUBX").toInt();